# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample ToDo list server."
  version: "1.0.0"
  title: "ToDo List API"
  termsOfService: "https://kotoblog.pp.ua/terms"
  contact:
    email: "spolischook@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "api.todo-list.kotoblog.pp.ua"
basePath: "/"
tags:
- name: "Task"
  description: "Everything about todo tasks"
  externalDocs:
    description: "Find out more"
    url: "https://en.wikipedia.org/wiki/Wikipedia:To-do_list"
schemes:
- "https"
paths:
  /tasks:
    get:
      tags:
        - "Task"
      summary: "Get list of tasks"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "owner"
          in: "header"
          required: true
          type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Task"
        400:
          description: "Miss header with tasks owner"
    post:
      tags:
        - "Task"
      summary: "Create new Task"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "owner"
          in: "header"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Task for create"
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                example: "ai1q755jt7tys1p"
      responses:
        201:
          description: "Success"
          schema:
            $ref: "#/definitions/Task"
        400:
          description: "Miss header with tasks owner"

  /tasks/{id}:
    get:
      tags:
        - "Task"
      summary: "Get one task"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "owner"
          in: "header"
          required: true
          type: "string"
        - name: "id"
          in: "path"
          description: "ID of task to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Task"
        400:
          description: "Miss header with tasks owner"
        404:
          description: "Task not found"
    post:
      tags:
        - "Task"
      summary: "Replace task content"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "owner"
          in: "header"
          required: true
          type: "string"
        - name: "id"
          in: "path"
          description: "ID of task to return"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          description: "Task for create"
          required: true
          schema:
            $ref: "#/definitions/TaskNew"
      responses:
        201:
          description: "Success"
          schema:
            $ref: "#/definitions/Task"
        400:
          description: "Miss header with tasks owner"
        404:
          description: "Task not found"
    patch:
      tags:
        - "Task"
      summary: "Edit task"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "owner"
          in: "header"
          required: true
          type: "string"
        - name: "id"
          in: "path"
          description: "ID of task to return"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          description: "Fields for edit"
          required: true
          schema:
            $ref: "#/definitions/TaskNew"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Task"
        400:
          description: "Miss header with tasks owner"
        404:
          description: "Task not found"
      deprecated: true
    delete:
      tags:
        - "Task"
      summary: "Delete task"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "owner"
          in: "header"
          required: true
          type: "string"
        - name: "id"
          in: "path"
          description: "ID of task to return"
          required: true
          type: "string"
      responses:
        200:
          description: "Succesfully deleted"
        400:
          description: "Miss header with tasks owner"

definitions:
  Task:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "string"
        example: "ai1qeyzjt7pgiot"
      title:
        type: "string"
        example: "Make my Home Work"
      responsible:
        type: "string"
        example: "Serhii Polishchuk"
      dueDate:
        type: "string"
        example: "2019-03-13 20:31:00"
      status:
        type: "string"
        description: "Task status"
        enum:
          - "new"
          - "in progress"
          - "done"
  TaskNew:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      title:
        type: "string"
        example: "Make my Home Work"
      responsible:
        type: "string"
        example: "Serhii Polishchuk"
      dueDate:
        type: "string"
        example: "2019-03-13 20:31:00"
      status:
        type: "string"
        description: "Task status"
        enum:
          - "new"
          - "in progress"
          - "done"